Fix Initial Captcha Requirement
There is a 'open the crate with the enclosed crowbar' style bug
in the current code which requires an admin password to set the
captcha keys, but requires captcha keys to create a new user.

This patch removes the captcha code from the sign-in page as long
as captcha keys are unset, which lets a deployment choose if/when
they will add captcha to the site, but, importantly, fixes the
bootstrap issue.

Testing:

Validated that this now works without a captcha key.

Note: As I don't have captcha creds, I've not tested that this
will work *with* captcha keys, so this is to be tested.